feat: Add examples of themed components and text styles
This commit adds examples of themed components and text styles to the end of the theme.dart file. The examples demonstrate how to use the AppTheme in a Flutter application and showcase various themed components, including buttons, switches, checkboxes, sliders, and text fields. It also showcases different text styles available in the AppTheme. These examples provide developers with a reference for implementing consistent styling throughout the application.

Recent user commits:
- refactor: Update import statement for theme in main.dart
- feat: Update expanded height and title in HomeScreen
- feat: Add input decoration theme for consistent styling
- update some colors
- add theme2
- Add script for generating test coverage report
- theme test version 2 test coverage 86.9 %
- theme tests. version 1
- revice comments