feat: Add Character class with Stats mixin
This commit adds a new class called Character in the character.dart file. The Character class uses the Stats mixin to add functionality for managing stats. It includes properties for name, slogan, and id, as well as a private field for encapsulation. The class also provides methods for toggling the favorite status and custom string representation. Additionally, it implements custom equality and hashCode for consistency. This new class enhances the RPG game by introducing a character with stats.

Recent user commits:
- chore: Update text styles in SampleScreen
- refactor: Update text styles in SampleScreen
- refactor: Rename HomeScreen to SampleScreen
- feat: Add examples of themed components and text styles
- refactor: Update import statement for theme in main.dart
- refactor: Update import statement for theme in main.dart
- feat: Update expanded height and title in HomeScreen
- feat: Add input decoration theme for consistent styling
- update some colors
- add theme2