feat: Update Character class to include skills
The Character class in character.dart has been updated to include a new field called `skills`, which is a set of Skill objects. This change allows characters to have associated skills and provides a way to update the skill set using the `updateSkills` method. This enhancement improves the RPG game by introducing a skill system for characters.

Recent user commits:
- chore: Update text styles in SampleScreen
- refactor: Update text styles in SampleScreen
- refactor: Rename HomeScreen to SampleScreen
- feat: Add examples of themed components and text styles
- refactor: Update import statement for theme in main.dart
- refactor: Update import statement for theme in main.dart
- feat: Update expanded height and title in HomeScreen
- feat: Add input decoration theme for consistent styling
- update some colors
- add theme2