refactor: Update CreateScreen to use SingleChildScrollView
The CreateScreen widget in create.dart has been updated to use SingleChildScrollView instead of ListView for the main content. This change improves the user experience by allowing the form to be scrollable when the keyboard is open, ensuring that all fields remain accessible. It enhances the usability of the character creation process and provides a more seamless interaction for the users.
ListView in Flutter uses lazy loading widgets to improve performance. What is happening here is that Flutter is not rendering your form field causing it to ignore your validation since validation is on the widget level. Since the widget is not visible, Flutter doesn't build it and therefore doesn't run the validation method you passed to it.

You can solve this by using a SingleChildScrollView with a Column. A Column doesn't lazy load and the widget will always be built whether or not the field is visible on screen, therefore allowing the Form widget to run the validation method on it.